LPS-66937 Move journal service data handlers to the proper package
